Created order placement and refactored into services

+++ src/test/java/com/tui/proof/ws/controller/ClientControllerTest.java
@@ -1,10 +1,13 @@
 package com.tui.proof.ws.controller;
 
-import com.tui.proof.dto.ClientDto;
+import com.tui.proof.dto.request.CreateClientRequest;
 import com.tui.proof.dto.ErrorDto;
 import com.tui.proof.model.Client;
 import com.tui.proof.repository.AddressRepository;
 import com.tui.proof.repository.ClientRepository;
+import com.tui.proof.repository.PilotesOrderRepository;
+import com.tui.proof.ws.controller.common.MockRequest;
+import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +20,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -28,18 +30,29 @@ public class ClientControllerTest {
     ClientRepository clientRepository;
     @Autowired
     AddressRepository addressRepository;
+    @Autowired
+    PilotesOrderRepository pilotesOrderRepository;
     @LocalServerPort
     private int port;
     @Autowired
     private TestRestTemplate restTemplate;
 
+    private MockRequest mockRequest = new MockRequest();
+
+    @After
+    public void truncateTables(){
+        pilotesOrderRepository.deleteAll();
+        addressRepository.deleteAll();
+        clientRepository.deleteAll();
+    }
+
     @Test
     public void createCustomerReturnsExpectedBody() {
-        ClientDto request = createValidRequest();
+        CreateClientRequest request = mockRequest.createValidRequest();
 
-        ResponseEntity<ClientDto> clientDtoResponseEntity = restTemplate.postForEntity("http://localhost:" + port + "/client/create", request, ClientDto.class);
+        ResponseEntity<CreateClientRequest> clientDtoResponseEntity = restTemplate.postForEntity("http://localhost:" + port + "/client/create", request, CreateClientRequest.class);
         assertEquals(HttpStatus.CREATED, clientDtoResponseEntity.getStatusCode());
-        ClientDto responseBody = clientDtoResponseEntity.getBody();
+        CreateClientRequest responseBody = clientDtoResponseEntity.getBody();
         assertEquals(request.getName(), responseBody.getName());
         assertEquals(request.getLastName(), responseBody.getLastName());
         assertEquals(request.getTelephone(), responseBody.getTelephone());
@@ -48,19 +61,26 @@ public class ClientControllerTest {
 
     @Test
     public void createCustomerInsertClientInDb() {
-        ClientDto validRequest = createValidRequest();
-        restTemplate.postForEntity("http://localhost:" + port + "/client/create", validRequest, ClientDto.class);
-        Client client = clientRepository.findClientByClientId(1);
-        assertEquals(validRequest.getName(), client.getFirstName());
-        assertEquals(validRequest.getLastName(), client.getLastName());
-        assertEquals(validRequest.getTelephone(), client.getTelephone());
+        CreateClientRequest validRequest = mockRequest.createValidRequest();
+        restTemplate.postForEntity("http://localhost:" + port + "/client/create", validRequest, CreateClientRequest.class);
+        List<Client> client = clientRepository.findAll();
+        int size = client.size();
+        if( size !=  1){
+            fail("expect the test databsae to be clean. Clients must be 1");
+        }
+
+        Client clienteCreated = client.get(0);
+
+        assertEquals(validRequest.getName(), clienteCreated.getFirstName());
+        assertEquals(validRequest.getLastName(), clienteCreated.getLastName());
+        assertEquals(validRequest.getTelephone(), clienteCreated.getTelephone());
         addressRepository.deleteAll();
         clientRepository.deleteAll();
     }
 
     @Test
     public void createCustomerValidation() {
-        ClientDto request = new ClientDto();
+        CreateClientRequest request = new CreateClientRequest();
         request.setLastName("");
         request.setTelephone("32199999999999999999999321321");
 
@@ -78,16 +98,4 @@ public class ClientControllerTest {
 
     }
 
-
-    private ClientDto createValidRequest() {
-        ClientDto request = new ClientDto();
-        request.setName("name");
-        request.setLastName("lastName");
-        request.setTelephone("321321321");
-        request.setStreet("Via 123");
-        request.setCity("Lecce");
-        request.setPostcode("73100");
-        request.setCountry("Italy");
-        return request;
-    }
 }
